+++ src/apis/goods.ts
@@ -45,3 +45,15 @@ export const getQueryKey = (params: SearchParams) => {
 export const getQueryFunction = (params: SearchParams) => {
   return () => getUsedGoods(params);
 };
+
+export const getUsedGoodDetail = async (id: string) => {
+  const { data, error } = await supabase
+    .from('used_item')
+    .select(
+      `*, profiles ( * ), main_category ( name ), sub_category ( name ), used_item_wish ( count ), chat_list ( count )`
+    )
+    .eq('id', id)
+    .single();
+
+  return data;
+};
